requirements.txt
streamlit
tensorflow-cpu==2.15.0
opencv-python-headless
pillow
numpy